fix: change analytics histogram data to double instead of long

+++ gravitee-apim-repository/gravitee-apim-repository-api/src/main/java/io/gravitee/repository/log/v4/model/analytics/HistogramAggregate.java
@@ -21,5 +21,5 @@ import java.util.Map;
 public interface HistogramAggregate {
     record Counts(String field, String name, Map<String, List<Long>> counts) implements HistogramAggregate {}
 
-    record Metric(String field, String name, List<Long> values) implements HistogramAggregate {}
+    record Metric(String field, String name, List<Double> values) implements HistogramAggregate {}
 }
